add book and navigation moules , array functions
