+++ index.html
@@ -6,7 +6,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="index.css">
-    <script defer src="index.js"></script>
+    <script defer type="module" src="index.js"></script>
     <title>Document</title>
 </head>
 
@@ -53,5 +53,4 @@
     </footer>
 
 </body>
-
 </html>
